Joystick input is correctly translated to targetRot and motorVal.

+++ src/pilot.h
@@ -8,7 +8,7 @@ class Pilot {
     int serRead;
     int serInput[PACKETSIZE];
     float axisVal[PACKETSIZE];
-    int mapLower, mapUpper;
+    float mapLower, mapUpper;
 //  double dir;   // Direction in radians
     bool okayToFly;
     long numGoodComm;
